+++ cmd/traffic/cmd/manager/managerutil/envconfig.go
@@ -28,6 +28,7 @@ type Env struct {
 
 	PodCIDRStrategy string `env:"POD_CIDR_STRATEGY,default=auto"`
 	PodCIDRs        string `env:"POD_CIDRS,default="`
+	PodIP           string `env:"TELEPRESENCE_MANAGER_POD_IP,default="`
 }
 
 type envKey struct{}